Avoid visual corruption if the game runs too fast.
If we run too fast while not in timedemo mode the game will
return immediately rather than rendering a frame.

If we go ahead and return to GLSurfaceView, it will swap the GL buffer
without having drawn anything into it, which will lead to garbage
being displayed.

The work-around is to detect that the game has decided to not render
anything, sleep for a ms and then try again.

Change-Id: Iba9d759547ddbc30534db963f5c47011038cc246